remove echo to add bashrc

+++ install.sh
@@ -41,27 +41,43 @@ sudo add-apt-repository -y ppa:apt-fast/stable
 sudo apt install -y apt-fast
 
 sudo apt-fast update
-sudo apt-fast install -y google-chrome-stable rpi-imager ibus-mozc mozc-utils-gui curl apt-transport-https git kicad clang clangd clang-format build-essential warp-terminal libreoffice docker-ce docker-ce-cli containerd.io blender obs-studio openvpn3 network-manager-l2tp network-manager-l2tp-gnome xsel nodejs npm python3-pip exuberant-ctags ripgrep eza make libssl-dev zlib1g-dev libbz2-dev libreadline-dev libsqlite3-dev llvm libncursesw5-dev xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev lldb libxslt1-dev autoconf bison libyaml-dev libncurses5-dev libgdbm-dev libdb-dev uuid-dev fzf htop net-tools trash-cli nmap openssh-server bacula-console-qt discord blueman apt-fast code direnv cargo sshpass
+sudo apt-fast install -y \
+  google-chrome-stable \
+  rpi-imager ibus-mozc mozc-utils-gui \
+  curl apt-transport-https git kicad clang \
+  clangd clang-format build-essential warp-terminal \
+  libreoffice docker-ce docker-ce-cli containerd.io blender \
+  obs-studio openvpn3 network-manager-l2tp \
+  network-manager-l2tp-gnome xsel nodejs npm python3-pip \
+  exuberant-ctags ripgrep eza make libssl-dev zlib1g-dev \
+  libbz2-dev libreadline-dev libsqlite3-dev llvm libncursesw5-dev \
+  xz-utils tk-dev libxml2-dev libxmlsec1-dev libffi-dev liblzma-dev \
+  lldb libxslt1-dev autoconf bison libyaml-dev libncurses5-dev \
+  libgdbm-dev libdb-dev uuid-dev fzf htop net-tools trash-cli nmap \
+  openssh-server bacula-console-qt discord blueman apt-fast code \
+  direnv cargo sshpass
 sudo apt-fast -y install docker-compose-plugin
 sudo apt-fast update
 
 sudo apt install -y tigervnc-viewer
 
-curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
-sudo rm -rf /opt/nvim
-sudo tar -C /opt -xzf nvim-linux64.tar.gz
-rm -f nvim-linux64.tar.gz
+if [ ! -d /opt/nvim ]; then
+  curl -LO https://github.com/neovim/neovim/releases/latest/download/nvim-linux64.tar.gz
+  sudo rm -rf /opt/nvim
+  sudo tar -C /opt -xzf nvim-linux64.tar.gz
+  rm -f nvim-linux64.tar.gz
+fi
 # add to bashrc
 # echo "export PATH="$PATH:/opt/nvim-linux64/bin"" >> ~/.bashrc
-echo "export PATH=\"\$PATH:/opt/nvim-linux64/bin\"" >>~/.my-bashrc
+# echo "export PATH=\"\$PATH:/opt/nvim-linux64/bin\"" >>~/.my-bashrc
 git config --global user.name rotarymars
 git config --global user.email s.goto2050@gmail.com
 git config --global commit.gpgsign true
 git config --global gpg.format ssh
-echo "source ~/.my-bashrc" >>~/.bashrc
-echo "alias pbcopy='xsel --clipboard --input'" >>~/.my-bashrc
-echo "alias g='git'" >>~/.my-bashrc
-echo "alias ls='eza --icons'" >>~/.my-bashrc
+# echo "source ~/.my-bashrc" >>~/.bashrc
+# echo "alias pbcopy='xsel --clipboard --input'" >>~/.my-bashrc
+# echo "alias g='git'" >>~/.my-bashrc
+# echo "alias ls='eza --icons'" >>~/.my-bashrc
 # git config --global user.signingkey ~/projects/ssh-keys/signing-keys.pub
 
 if [ ! -d ~/.config ]; then
@@ -87,15 +103,15 @@ fi
 sudo npm install n -g
 
 sudo n stable
-sudo apt purge -y nodejs npm
+# sudo apt purge -y nodejs npm
 sudo apt autoremove -y
 pip install jedi
 pip install git+https://github.com/rotarymars/oj.git
-echo "export PATH=\"\$PATH:/home/rotarymars/.local/bin\"" >>~/.my-bashrc
+# echo "export PATH=\"\$PATH:/home/rotarymars/.local/bin\"" >>~/.my-bashrc
 sudo npm install -g atcoder-cli
-echo "alias compile='clang++ -g -std=c++2b'" >>~/.my-bashrc
-echo "alias accnew='acc new --template kyopuro'" >>~/.my-bashrc
-echo "alias accadd='acc add --template kyopuro'" >>~/.my-bashrc
+# echo "alias compile='clang++ -g -std=c++2b'" >>~/.my-bashrc
+# echo "alias accnew='acc new --template kyopuro'" >>~/.my-bashrc
+# echo "alias accadd='acc add --template kyopuro'" >>~/.my-bashrc
 
 # open /etc/group
 # sudo vi /etc/group
@@ -105,8 +121,8 @@ echo "alias accadd='acc add --template kyopuro'" >>~/.my-bashrc
 
 if [ ! -d ~/.asdf ]; then
   git clone https://github.com/asdf-vm/asdf.git ~/.asdf --branch v0.13.1
-  echo ". \"\$HOME/.asdf/asdf.sh\"" >>~/.my-bashrc
-  echo ". \"\$HOME/.asdf/completions/asdf.bash\"" >>~/.my-bashrc
+  # echo ". \"\$HOME/.asdf/asdf.sh\"" >>~/.my-bashrc
+  # echo ". \"\$HOME/.asdf/completions/asdf.bash\"" >>~/.my-bashrc
   asdf plugin add python
   asdf plugin add ruby
   asdf install python latest
@@ -121,7 +137,7 @@ fi
 
 sudo gpasswd -a rotarymars docker
 
-echo "alias rm='trash-put'" >>~/.my-bashrc
+# echo "alias rm='trash-put'" >>~/.my-bashrc
 git config --global core.editor 'code --wait'
 
 # replace cap
@@ -131,3 +147,10 @@ TEMPLATE_PATH=$(dirname $0)/templates
 cp ${TEMPLATE_PATH}/.ssh/config ~/.ssh/config
 cp ${TEMPLATE_PATH}/.my-bashrc ~/.my-bashrc
 
+# add to load my-bashrc if necessary
+grep .my-bashrc ~/.bashrc >> /dev/null
+EC=$?
+if [ $EC -ne 0 ]; then
+  echo set .my-bashrc to .bashrc
+  echo "source ~/.my-bashrc" >> ~/.bashrc
+fi
